requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
google-api-python-client==2.125.0
google-auth==2.29.0
